+++ example/logging-demo/instance-logging/Cargo.toml
@@ -11,5 +11,6 @@ crate-type = [ "cdylib" ]
 
 [dependencies]
 invoke-witc = { path = "../../../bindings/rust/invoke-witc" }
+witc-abi = { path = "../../../bindings/rust/witc-abi" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "*"
