+++ example/traffic-lights/instance-lights/Cargo.toml
@@ -13,3 +13,4 @@ crate-type = [ "cdylib" ]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "*"
 invoke-witc = { path = "../../../bindings/rust/invoke-witc" }
+witc-abi = { path = "../../../bindings/rust/witc-abi" }
