+++ example/traffic-lights/instance-lights/src/lib.rs
@@ -1,5 +1,6 @@
 #![feature(wasm_abi)]
 use serde::{Deserialize, Serialize};
+use witc_abi::instance::*;
 invoke_witc::wit_instance!(export("traffic-lights.wit"));
 
 fn toggle(status: light) -> light {
